Reading from database stage fully complete.
LifeObjective: New method to return Type as an integer.
Activity: Date storage changed from GregorianCalendar to Timestamp.
DBConnector: Added a function to close the connection.
Query: Fixed getData function to reflect Activity constructor and to be
simpler.
Main: Database connection, simple query and writing the data to console.
Most of it should be temporary.