Multiple dates and places for the same activity.
Main: Minor changes, just displays the second date and time too.
Activity: Changed member variable to a list, modified the setter and
getter functions accordingly, added a function to add a single
PlaceAndTime object to the list.
Query: Modified getData function, so it will not make a duplicate
Activity object when multiple dates and places are in the database for
the same activity, only adds to its list.