Test driven changes
Query: GetData changed due to Activity.PlaceAndTime changes. DeleteData
will not try to delete anything from the DB if it doesn't exist.
StoreData single activity function will return if the activity was
already in the DB or not. The list type function will return a list with
the data already in the DB.
Activity: New variable added to nested class: duration. Removed and
added some constructors from the core and nested class. AddPlaceAndTime
modified to keep "free date" if chosen. Added function returning
"sql-compatible" Strings for place, duration and timeAndDate.
Main: Minor changes due to changes to addPlaceAndTime and the new
variable.