Schedule Id added
Activity: New byte variable scheduleId added. It will help differentiate
between weekly/daily and one-time "happenings"(can be just a part of an
activity). This is a number between 0 and 127, 0 for unscheduled and 127
for daily. Each power of 2 contained is a day of the week from Monday
(2^0) to Sunday (2^6). The sum gives us the days it refers to.
Query: storeData and getData changed to query and store the new
variable. SQL database is modified accordingly.
Main: "Cosmetic" changes to printStuff and prints the new variable too.